pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "xclip"
version = "0.1.0"
authors = [
  { name="Elias Kempf", email="kempfe@cs.uni-freiburg.de" },
  { name="Simon Schrodi", email="schrodi@cs.uni-freiburg.de" }
]
description = "Code of our paper 'When and How Does CLIP Enable Domain and Compositional Generalization?'"
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]

dependencies = [
    "torch==2.4.1",
    "torchvision==0.19.1",
    "lightning==2.4.0",
    "numpy==1.26.4",
    "tqdm==4.66.2",
    "tensorboard==2.16.2",
    "timm==1.0.7",
    "transformers==4.44.2",
    "datasets==2.19.2",
    "dill==0.3.8",
    "pillow==10.3.0",
    "transformer-lens==2.7.0",
    "scikit-learn==1.5.1",
    "matplotlib==3.9.0",
    "seaborn==0.13.2",
    "tueplots==0.0.17",
    "plotly==5.22.0",
    "pandas==2.2.2",
    "tabulate==0.9.0",
    "spacy==3.7.5",
    "textacy==0.13.0",
    "nltk==3.8.1",
    "ftfy==6.2.0",
    "jaxtyping==0.2.34",
    "graphviz==0.20.3",
    "nnsight==0.3.7",
    "networkx==3.4.2",
    "umap-learn==0.5.6",
    "torch-pca==1.0.0",
    "braceexpand==0.1.7",
    "webdataset==0.2.86",
    "sentencepiece==0.2.0",
]

[project.optional-dependencies]
dev = [
    "torch-tb-profiler",
    "pre-commit",
    "nvitop",
    "ruff==0.9.10"
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.ruff]
line-length = 120
indent-width = 4
target-version = "py310"
exclude = ["deps"]

[tool.ruff.lint]
select = ["E4", "E7", "E9", "F", "I"]

[tool.ruff.format]
quote-style = "single"
